Title: 합과 평균, Time: 150ms, Memory: 11MB, Status: Passed - CodeTree

+++ 231214/합과 평균/sum-and-avg.java
@@ -9,7 +9,7 @@ public class Main {
 
 
 
-        System.out.print(n+a + " " + (n+a)/2);
+        System.out.print((int)(n+a) + " " + (n+a)/2);
 
 
     }
